requirements.txt
pyinstaller>=6.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
PyQt5>=5.15.0

